Number of chocolate piles program in python.
Problem : Let's say you are making n chocolate piles in a row. The first pile has n chocolates. The number of chocolates in each pile depends on the first pile. If the first pile has an even number of chocolates, then all piles must have an even number of chocolates. Similarly, if the first pile has an odd number of chocolates, then all piles must have an odd number of chocolates. Each pile must have more chocolates than the previous pile, but as few as possible. Write a Python program to print the number of chocolates in each pile.